Moving the Close tab event from ViewLoadedParams to WorkspaceDependencyView call
